clean-up of select
