+++ src/components/forms/index.js
@@ -1,13 +1,11 @@
 import Checkbox from './Checkbox'
 import Radio from './Radio'
-import Select from './Select'
 import Switch from './Switch'
 import TextField from './TextField'
 
 export default {
   Checkbox,
   Radio,
-  Select,
   Switch,
   TextField
 }
