+++ src/components/selects/index.js
@@ -1,0 +1,5 @@
+import Select from './Select'
+
+export default {
+  Select
+}
